Refactor: Improve mobile work order card layout
Implement mobile-friendly work order card layout using `useIsMobile` hook, reorganizing information, optimizing spacing, and improving touch targets.
